:name_badge: Rename for UniFlow.Signal
