+++ Assets/Scripts/TimelineSignal/Begin.cs
@@ -1,4 +1,4 @@
-namespace UniFlow.Signal
+namespace UniFlow.TimelineSignal
 {
     public class Begin : SignalAssetBase
     {
