+++ Assets/Scripts/TimelineSignal/SignalAssetBase.cs
@@ -1,6 +1,6 @@
 using UnityEngine.Timeline;
 
-namespace UniFlow.Signal
+namespace UniFlow.TimelineSignal
 {
     public class SignalAssetBase : SignalAsset
     {
